Agents/requirements.txt
pandas
qrcode[pil]
requests
python-dotenv

